backend/requirements.txt
Flask==3.0.0
alembic==1.13.1
pytest==7.4.4
pytest-cov==4.1.0
flake8==7.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
